+++ rust/cli/CHANGELOG.md
@@ -1,10 +1,22 @@
 # Changelog
 
-## 0.1.4-dev
+## 1.0.0
+
+### Patch
+
+- Change description
+- Update dependencies
+
+## 0.1.4
 
 ### Minor
 
 - Use true colors when available
+
+### Patch
+
+- Dissociate repository from published content (see `publish.sh` script)
+- Remove `package.metadata.deb`
 
 ## 0.1.3